requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

torch>=2.2
soundfile>=0.12.1
numpy>=1.26

openai-whisper==20231117

# GigaAM (если используете его)
# gigaam==<ваша_версия>
hydra-core==1.3.*
numpy==2.*
omegaconf==2.3.*
onnx==1.19.*
onnxruntime==1.23.*
sentencepiece
torch>=2.5,<2.9
torchaudio>=2.5,<2.9
tqdm    